# openapi.json
{"openapi":"3.0.2","info":{"title":"app","description":"My new FastAPI project","version":"0.1.0"},"paths":{"/health_check/ping_application":{"get":{"tags":["Application Health"],"summary":"Ping Application","operationId":"ping_application_health_check_ping_application_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PingResponse"}}}}}}},"/health_check/ping_database":{"get":{"tags":["Application Health"],"summary":"Ping Database","operationId":"ping_database_health_check_ping_database_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PingResponse"}}}},"500":{"description":"Database isn't working"}}}},"/groups":{"get":{"tags":["Groups"],"summary":"Get Groups","operationId":"get_groups_groups_get","parameters":[{"required":false,"schema":{"title":"Query","type":"string"},"name":"query","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Groups Groups Get","type":"array","items":{"$ref":"#/components/schemas/GroupForList"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Groups"],"summary":"Create Group","operationId":"create_group_groups_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupCreate"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupCreateResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/groups/{group_id}":{"get":{"tags":["Groups"],"summary":"Get Group","operationId":"get_group_groups__group_id__get","parameters":[{"required":true,"schema":{"title":"Group Id","type":"integer"},"name":"group_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupInfo"}}}},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Groups"],"summary":"Update Group","operationId":"update_group_groups__group_id__put","parameters":[{"required":true,"schema":{"title":"Group Id","type":"integer"},"name":"group_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/GroupUpdate"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Groups"],"summary":"Delete Group","operationId":"delete_group_groups__group_id__delete","parameters":[{"required":true,"schema":{"title":"Group Id","type":"integer"},"name":"group_id","in":"path"}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/students":{"get":{"tags":["Students"],"summary":"Get Student","operationId":"get_student_students_get","parameters":[{"required":false,"schema":{"title":"Query","type":"string"},"name":"query","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"title":"Response Get Student Students Get","type":"array","items":{"$ref":"#/components/schemas/StudentInfo"}}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["Students"],"summary":"Create Student","operationId":"create_student_students_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentInfo"}}}},"201":{"description":"Created"},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/students/{student_id}":{"get":{"tags":["Students"],"summary":"Get Student","operationId":"get_student_students__student_id__get","parameters":[{"required":true,"schema":{"title":"Student Id","type":"integer"},"name":"student_id","in":"path"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentInfo"}}}},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["Students"],"summary":"Update Student","operationId":"update_student_students__student_id__put","parameters":[{"required":true,"schema":{"title":"Student Id","type":"integer"},"name":"student_id","in":"path"}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/StudentUpdate"}}},"required":true},"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["Students"],"summary":"Delete Student","operationId":"delete_student_students__student_id__delete","parameters":[{"required":true,"schema":{"title":"Student Id","type":"integer"},"name":"student_id","in":"path"}],"responses":{"204":{"description":"Successful Response"},"404":{"description":"Not Found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"GroupCreate":{"title":"GroupCreate","required":["name"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"parent_id":{"title":"Parent Id","type":"integer"}}},"GroupCreateResponse":{"title":"GroupCreateResponse","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"parent_id":{"title":"Parent Id","type":"integer"}}},"GroupForList":{"title":"GroupForList","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"subGroups":{"title":"Subgroups","type":"array","items":{"$ref":"#/components/schemas/GroupInfo"}}}},"GroupInfo":{"title":"GroupInfo","required":["id","name"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"}}},"GroupUpdate":{"title":"GroupUpdate","required":["name","id"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"parent_id":{"title":"Parent Id","type":"integer"},"id":{"title":"Id","type":"integer"}}},"HTTPValidationError":{"title":"ErrorResponse","required":["detail"],"type":"object","properties":{"detail":{"title":"Detail","type":"object","example":[{"field":"quantity","message":"value is not a valid integer"}]}}},"PingResponse":{"title":"PingResponse","required":["message"],"type":"object","properties":{"message":{"title":"Message","type":"string"}}},"StudentCreate":{"title":"StudentCreate","required":["name","group_id"],"type":"object","properties":{"name":{"title":"Name","type":"string"},"email":{"title":"Email","type":"string","format":"email"},"group_id":{"title":"Group Id","type":"integer"}}},"StudentInfo":{"title":"StudentInfo","required":["id","name","group_id"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"group_id":{"title":"Group Id","type":"integer"}}},"StudentUpdate":{"title":"StudentUpdate","required":["id","name","group_id"],"type":"object","properties":{"id":{"title":"Id","type":"integer"},"name":{"title":"Name","type":"string"},"group_id":{"title":"Group Id","type":"integer"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}},"tags":[{"name":"ya_profi","description":"My new FastAPI project"}]}